+++ .gitignore
@@ -24,3 +24,9 @@ __pycache__
 *.ipynb_checkpoints
 *.ipynb
 wandb
+tests
+dist
+*.egg-info
+build
+*.egg
+*.whl
