--- easylm/__init__.py
@@ -1,6 +1,0 @@
-# src/__init__.py
-from .config import *
-from .tokenizer import *
-from .nn import *
-from .model import *
-from .data import *
