--- easylm/data/__init__.py
@@ -1,2 +1,0 @@
-from easylm.data.dataset import *
-from easylm.data.dataloader import *
