--- easylm/model/__init__.py
@@ -1,4 +1,0 @@
-# src/model/__init__.py
-from .gpt import *
-from .llama import *
-from .bert import *
