+++ src/langtrain/nn/_nn.py
@@ -3,8 +3,7 @@ import torch
 from torch import nn
 from typing import Tuple, Union
 import torch.nn.functional as F
-
-from easylm.config import VITConfig
+from langtrain.config._config import VITConfig
 
 class Linear(nn.Module):
     def __init__(self, in_features: int, out_features: int, bias: bool = True) -> None:
