+++ src/langtrain/tokenizer/__init__.py
@@ -1,0 +1,1 @@
+from ._sentencepiece import Tokenizer
