+++ src/langtrain/trainer/__init__.py
@@ -1,0 +1,1 @@
+from ._trainer import Trainer
