+++ src/langtrain/utils/_utils.py
@@ -1,7 +1,8 @@
-from typing import List, NamedTuple, Union
+from typing import List, NamedTuple, Tuple, Union
 import numpy as np
 import torch
 import torch.nn.functional as F
+from torch.nn.utils.rnn import pad_sequence
 
 
 def trainable_parameters(model: torch.nn.Module)->str:
@@ -23,3 +24,10 @@ class MaskedModelOutput(NamedTuple):
     logits: Union[torch.Tensor, None] = None
     hidden_states: Union[torch.Tensor, None] = None
     pooler_output: Union[torch.Tensor, None] = None
+
+
+def collate_fn(batch: List[Tuple[torch.Tensor, torch.Tensor]]) -> Tuple[torch.Tensor, torch.Tensor]:
+    inputs, targets = zip(*batch)  # Unpack the batch into inputs and targets
+    inputs = pad_sequence(inputs, batch_first=True, padding_value=0)
+    targets = pad_sequence(targets, batch_first=True, padding_value=-100)
+    return inputs, targets
